requirements.txt
numpy==1.23.2
pycairo==1.21.0
PyGObject==3.42.2
pyserial==3.5
